apps/ml/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
lightgbm==4.5.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
redis==5.0.8
pydantic==2.9.2
python-dotenv==1.0.1
